tous marche CRUD

+++ src/Service/SqlService.php
@@ -70,6 +70,7 @@ class SqlService
             $pre =  $this->connection->prepare($request);
 
             $pre->execute($param);
+            
         } catch (PDOException $e) {
             echo 'Error  ' . $e->getMessage();
         }
@@ -85,7 +86,19 @@ class SqlService
 
     
     public function remove(string $nameTable, int $id)
-    {
+    {  
+       $users =  SqlService::getInstance()->fetch($nameTable);
+        $userCourant = null;
+        foreach ($users as $user) {
+          if ($user->id == $id) {
+            $userCourant = $user;
+          }
+        }
+        if ($userCourant == null) {
+         echo 'Cet utilisateur n\'existe pas, impossible de le supprimer !!!!';
+         die();
+        }
+      
 
         $request = "DELETE FROM  " . $nameTable . " WHERE id= :id";
